cross-merging fork to branch from Iceman's repository

+++ common/cmd.h
@@ -33,12 +33,14 @@
 #ifndef _PROXMARK_CMD_H_
 #define _PROXMARK_CMD_H_
 
-#include <common.h>
-#include <usb_cmd.h>
+#include "common.h"
+#include "usb_cmd.h"
 #include "usb_cdc.h"
+#include "usart.h"
+#include "proxmark3.h"
 
 bool cmd_receive(UsbCommand* cmd);
-bool cmd_send(uint32_t cmd, uint32_t arg0, uint32_t arg1, uint32_t arg2, void* data, size_t len);
+bool cmd_send(uint64_t cmd, uint64_t arg0, uint64_t arg1, uint64_t arg2, void* data, size_t len);
 
 #endif // _PROXMARK_CMD_H_
